In Node, you never see path strings concatenated together. Nope, to keep things portable between Unix and Windows, you're supposed to call path.resolve() - or a similar function path.join() - to concatenate the path for you.
You can pass as many arguments as you want: the args are built into one big path string.

 ./node_modules/.bin/webpack
Ah! An error!

  path is not defined

Listen up! This is important: webpack.config.js is executed by Node... and this code will never run in our browser. So since this is a pure Node script, we can use any of the core libraries that you normally have access to, like the path module.