Re-run webpack, but this time with a --watch: ./node_modules/.bin/webpack --watch
It still builds the assets... but then waits. It's now constantly watching all your files for changes.

This is nothing earth-shattering... it's just a tool we need... and Webpack provides it without any setup. In the terminal, you can even see that it re-dumped our file.

There is one caveat with watch: and we'll see it a few times. If you make a change to the webpack.config.js file itself, you will need to manually restart webpack. Webpack watch does not detect changes to itself.