Right now, RepLogApp only works because jQuery is available globally:
 public/js/RepLogApp.js
(function(window, $, Routing, swal) {
})(window, jQuery, Routing, swal);

We have a self-executing function, which expects jQuery to be a global variable by the time this file is loaded. That works because, in our base layout - app/Resources/views/base.html.twig - we include jQuery.

This gives us two global variables - $ and jQuery - which we use everywhere else.

And this is the big, huge, classic, horrifying problem we want to fix: when we write JavaScript, our code is not self-contained. Nope, we need to very carefully make sure to remember to include any dependent script tags before loading our JavaScript. If you forget to add a script tag or add it in the wrong order... bam! Your user sees an error. And you see your face hitting your palm.

This is a terrible and unprofessional way to live. But with Webpack, we can finally fix this. Instead of crossing your fingers and hoping that jQuery is included on this page, in RepLogApp.js, we will require it.