Try it: head back to LiftStuff and refresh. Yes! It still loads!
But when we hit 'Delete'... woh... something isn't right: it looks like I designed this... which you do not want. And... I can't hit 'Cancel'. What's going on here?

I'll open up my debugger. Huh... no errors. This is actually a CSS problem. When we installed sweetalert2, it installed version 6.6.10.

And when you use SweetAlert, you actually need to make sure you include its JavaScript on the page... and also the SweetAlert CSS file. We're loading the CSS file from a CDN... but at a different version: 6.1.0