Since RepLogApp is being loaded by webpack. it's already a module.
Now, if anything requires this file, they will get the RepLogApp class. And, we are no longer modifying anything in the global scope.

And as soon as we try that, our app is super broken! Thanks Ryan!

  RepLogApp is not defined

Yes! This makes sense: in our template, we're still trying to reference the now,

non-existent - global variable RepLogApp

How do we fix this? By fully modularizing our code and removing all JavaScript from our templates.