Our simple site actually has two pages: this main 'lift stuff' page, and also a login page... which has just a little bit of its own JavaScript: a toggleable message and an error if you get crazy and type a super-long username.
Open up the template: templates/security/login.html.twig. Yep, it has a script tag for assets/js/login.js: a traditional, non-webpacked file.

And... that's a bummer! I want Webpack to also process this file so I can remove the self-executing function and require jQuery instead of hoping it's included on the page:
23 lines  public/js/login.js
(function(window, $) {
})(window, jQuery);