When we refresh, we do not get warm, fuzzy feelings: we get a cold and not-so-fuzzy error:   $ .tooltip() is not a function
Bootstrap's job is to modify jQuery and add more functions to it, like tooltip. But, in layout.js, we require a different, fresh, unmodified jQuery and then try to call .tooltip() on it!

That doesn't work because - when the Bootstrap code is loaded - it modifies the jQuery that's loaded in our layout: it modifies the global jQuery variable. In layout.js, we're not using the global jQuery variable: we're requiring our own jQuery... which has not been modified by Bootstrap.

The fix? We already know it: we need to once again handle all our dependencies inside layout.js. I mean, if we need bootstrap functions inside of this file, then we need to require bootstrap here.