Even with Babel, this is still needed. Babel transpiles our new JavaScript syntax into old JavaScript... but it doesn't add missing features, like the Promise object.
So yes... we still need something to fill in the missing Promise object... but I do not want to keep using these global script tags.

This library will give us the Promise polyfill and actually, all Polyfills. Use it inside of layout.js - so that we get the polyfill on every page: require @babel-polyfill.