We're using css-loader and style-loader. When you activate loaders with the inline syntax we saw earlier, you read the loaders from right to left. Here, it's similar: read the loaders from bottom to top.
So, this says: when we require a file ending in .css, first send its contents through css-loader - which converts it to that JavaScript object - and then send that through style-loader.

And what does style-loader do? Let's find out! Go back to your watch tab and restart webpack: ./node_modules/.bin/webpack --watch
Ok, find your browser and refresh.

What? It works! Our CSS is back... even though we don't have a link tag on the page! The console logs an empty value.

But if you inspect the page and look in the head tag, there's a surprise: a new style element with all our CSS. Yep! the style-loader's job is to package the CSS inside login.js, along with some extra JavaScript that injects that CSS onto the page in a style tag. It's incredible.

But, it does have one downside. Refresh and watch closely: the page is ugly for just a moment, before the JavaScript loads. Then, it looks fine. But, for production, that's a problem. Yep, style-loader is a development-only tool. But, I promise to show you a proper, great solution for production later.

For now, celebrate! We can require CSS from our JavaScript! We are truly creating self-contained JavaScript apps.