That should just work!
Refresh! And as promised... woh! Nevermind! It does not work... it looks terrible! What happened?

Our console shows an error:

   Module parse failed, dummbell-mini.png, unexpected character

Huh. Over in the terminal, the watch script has a similar message... and it looks like it happens when Webpack reads main.css.

Huh. Over in the terminal, the watch script has a similar message... and it looks like it happens when Webpack reads main.css.

Hmm. Open main.css. Ah! There's the image: it's a background image inside our CSS!

 78 lines  web/assets/css/main.css
.mini-dumbbell {
    background: url('../images/dumbbell-mini.png') center center  no-repeat;
}
This, is very interesting. When we tell Webpack to load a CSS file, it actually parses the background images and - basically - requires them! It does the same with fonts and also finds and requires any @import calls.

So... why is this failing? Well, just like with CSS, a .png file is not JavaScript... so Webpack has no idea what to do with it!