We don't need to worry about requiring a minified file because, eventually, Webpack will minify all our CSS for us.
What Happens when you Require a Module Name?

But then, how does this work? When we just say require('bootstrap')... what does that do? Does it look for an index.js file in that directory? Well, there isn't one, so there goes that theory.

The secret is inside of the package.json file for that package. It should have a key called main. Not all libraries have this key - some old, legacy libraries won't - but most new libraries will have this.

The main key tells Node which file to include when someone simply requires the module by name. When we say, require('bootstrap') it actually requires dist/js/npm.js.

file-loader for Fonts
Close package.json, and go refresh the page! Woh! That did not go well. In the console... ah! A familiar error:
Module parse failed for fonts/glyphicons-halflings-regular.eot. Unexpected character

Yep, Bootstrap's CSS is referencing a font file, and just like with images, Webpack has no idea what to do with it!

Back on the Asset Management section of Webpack's site, there's also a link for Loading Fonts. And, yea, it's the same exact thing as images.