Head back to the Webpack site and click back on 'Loading Images'. There's a link to the documentation for the file-loader(https://webpack.js.org/loaders/file-loader/).
It shows how to use the file-loader using the inline syntax, which is one of the reasons I wanted you to see it. It looks like there's an option called name... which should let us control the name of those output files. And, just like with Webpack's output, it has some special placeholders, like [name], [extension] and [hash].

To see this in action, in your terminal, stop webpack, and then clear out the build/ directory:

rm -rf web/build/*
Ok, come back Webpack!
./node_modules/.bin/webpack --watch
Yes! Now, the filenames are way prettier. We can quickly see which files are which because they begin with their original name. Only six character at the end are from the hash.