This package is just like the previous one... except it has Sass files instead of CSS. It does still contain the Bootstrap JavaScript. So to keep loading that, just change the require to bootstrap-sass.
Next, for the CSS require, we need to find the new path. Look inside node_modules/ for bootstrap-sass. Ok! It has an assets/stylesheets/_bootstrap.scss file! This is the key: it imports every piece of bootstrap.

Check out the terminal that's running Webpack. It had a ton of errors... but now it's happy again! Try out the site. Nice!