Up until now, to load CSS or Sass, we've required it from JavaScript. But, you can also use @import from inside CSS. Close the node_modules/ directory and open our main.scss.
Webpack parses these @import lines... so this is effectively identical to what we had before. Except... it doesn't work! Even PhpStorm is super angry! When you use require inside a JS file, Webpack knows to look for bootstrap-sass inside the node_modules/ directory. But when you use a CSS @import, that path is relative to this file.