To hint to Webpack that you want to treat this path like a module path, prefix it with ~ Let's do this for Font Awesome too: copy its path, remove the require line.
In theory... this should work just like before. Right? Right? Try it!

Oh no! It looks terrible! Find your terminal. Interesting:

  Can't resolve fonts/bootstrap/glyphicons-halflings-regular.svg

... and then it references our main.scss file. Huh. This is really odd.

What's really curious is that the file is actually referenced from deep inside bootstrap: we require _bootstrap.scss, it imports another _glyphicons.scss and it references the font files. Well, it uses some variables, but this code references that file.

And this worked a moment ago... so why would it break? All we did was change the way that we were require these files.

So... this is a little confusing. Without going into too much detail, when you @import a Sass file from another Sass file in a different directory, any paths wrapped in url() get totally messed up: Webpack, well really, css-loader, looks for those paths relative to the first Sass file. In other words, it looks for the font files relative to main.scss, not _glyphicons.scss