Now that we finally have Bootstrap's Sass working, we can celebrate! Inside _bootstrap.scss, the first thing it imports is an _variables.scss file. I'll hold Command and click to open that.
Awesome! This holds a bunch of variables that are used throughout the rest of the Sass files. Yep, this means we can override these to control colors and many other things.

Copy the $ brand-primary variable line. Then, at the top of our main.scss file, before we import Bootstrap, override that variable. Make sure to remove the default.

This variable is used to color some buttons and a lot of other things. Let's make it a little lighter: change the 6.5 to 2.5. It's not important, but the new color should be #3885C7

Back in the browser, inspect the 'I Lifted It! ' button. Yep! The current color is #337AB7. Now, refresh! Check it out. Yes! It's the slightly lighter #3885C7.

So this is really the reason why you might want to import Bootstrap as Sass, instead of CSS.