In _bootstrap.scss, comment out the glyphicons line.
We don't need to restart Webpack after this change, but, stop it temporarily so that we can clean out the build/ directory:

rm -rf web/build/*
Now, re-run webpack: ./node_modules/.bin/webpack --watch

Check it out! In web/build, we still have the Font Awesome fonts... but we do not have the glyphicon fonts anymore. That's awesome! We're not parsing the _glyphicons.scss file anymore, so Webpack never sees those font files and so never moves them. This is proof that our CSS file is now just a little bit smaller.