When I hover over the leaderboard, I want to show a JavaScript Bootstrap popover that gives us some info about how it works.
To get this working, all we need to do is write some JavaScript to find that element and activate the popover. Normally, we would put this in rep_log.js. In fact, that's the best place to put it. Well, there, or some file that it requires. The point is, now that our JavaScript has matured, 100% of your JS code should live in an external JavaScript file. You should have no JavaScript in your templates, other than the one script tag:
53 lines  app/Resources/views/lift/index.html.twig
{% block javascripts %}

    <script src='{{ asset('build/rep_log.js') }}'></script>

{% endblock %}
But, what if you're using Webpack on a legacy app... that still has a ton of JavaScript in your templates? You should refactor this. But, since that can be a big job, let's find out how we can still write JavaScript in a template... even though we shouldn't.

It should work, right? After all, rep_log.js requires jQuery, so we should be able to use it in our template.

Well, try it. Refresh the page. Uh oh... error!

  $ is not defined

This is coming from our template. What's going on? We are requiring jQuery!

We already know what's going on. When we require jquery, it detects that it is being used in a module environment. So it does not set a global jQuery or $ variable. Instead, it just returns the jQuery function, which we set to $. But that only affects this file.

And this is good! We want each module to live in isolation like this.