If you do need to expose a jQuery variable globally, you can. But remember, this is not ideal, and is just meant to support legacy code.
But, I have one tiny complication. In webpack.config.js, near the bottom, we added something called the webpack.ProvidePlugin():

webpack.config.js
module.exports = {
    plugins: [
        new webpack.ProvidePlugin({
            jQuery: 'jquery',
            $ : 'jquery'
        }),
    ],
};
Thanks to this, if any JavaScript relies on jQuery or $ to be available as a global variable, Webpack rewrites that code to require the jquery module properly. This does not make the jQuery or $ variables available globally, like in a template. Instead, it fixes code to not need that. If you do want to make these variables global, that's what we just did in rep_log.js:

 13 lines  assets/js/rep_log.js
// expose $ globally so I can use it in the template
// ... even though I should put all my code here!
window. $ = $;
In other words, these are solving two separate, but similar issues.

Here's the complication: sometimes, a third party library will try to reference jQuery globally not by just saying jQuery, but by saying window.jQuery... which is really the same thing. But, right now, the ProvidePlugin would not fix that situation. So we could still have some issues with some libraries.