Open up the base layout: app/Resources/views/base.html.twig. Scroll down to the bottom:
 103 lines  templates/base.html.twig
{% block javascripts %}
    <script src='{{ asset('bundles/fosjsrouting/js/router.js') }}'></script>
    <script src='{{ path('fos_js_routing_js', { callback: 'fos.Router.setData' }) }}'></script>
    <script src='{{ asset('build/layout.js') }}'></script>
{% endblock %}
In a perfect world, we would only have one JS file: layout.js. But, we still have two other script tags that are cramping my style. One loads router.js from FOSJsRoutingBundle, and the second populates that with the actual routes.

Thanks to these, there is a global Routing variable, which we use in RepLogApp.js to generate URLs: Routing.generate() and then the route name:

assets/js/Components/RepLogApp.js
class RepLogApp {
    _saveRepLog(data) {
        return new Promise((resolve, reject) => {
            const url = Routing.generate('rep_log_new');
        });
    }
}
This is the last place where we are using global variables... and I want to fix that!