The curious thing about the router.js file is that it lives in a PHP package. Instead of living in node_modules, it lives at vendor/friendsofsymfony/jsrouting-bundle/Resources/js/router.js. This is the exact file we're currently including in base.html.twig.
Phew! Long, gross-looking... but it makes sense. Try it. Refresh!

Uh oh... it did not like that:

  Type error: Cannot read property navigator of undefined

It's coming from router.js, line 8. Huh. If you look at this file, well, it's minified... so it's pretty confusing. This file is built using something called the Google Closure compiler... which basically means that it does not play nicely with require(). I hope this will be fixed in the library soon - there's a pull request open for it. But right now... well, there is no simple way to require() it.

 Tip

It is possible to require() the file... but it takes some work. See bit.ly/fosjsroutingbundle-require if you're curious.

So... yea... we need to keep these two script tags in our base layout for now. It's not ideal, but we'll survive.