When it comes to using require() and module.exports, well, we rock. So now... let's completely change things! Woohoo!
require() and module.exports basically come from Node.js. They're not part of ECMAScript - so not officially part of the JavaScript language. But, if you're in Node, you know that these are available to use.

Node did this because... well... there was no module system in ECMAScript. So they invented require() and module.exports to fit their needs. But, guess what! In ES6... there now is a module system! It uses two new keywords: import and export.

Webpack supports the original require() and module.exports and the ES6 syntax. But, import and export have a few subtle advantages... other than being newer and shinier.