If you read the documentation about the webpack-dev-server, you'll read about HMR: heavy metal rock. Or, actually, in this context, it means hot module replacement. HMR is amazing: it allows Webpack to update your page... without refreshing! But... you need to configure your system correctly to be able to do this. And, unless you do even more setup, it only works for CSS. You can make it work for JavaScript - especially if you're using a framework like React or Vue, but you won't get it out of the box.
Anyways, let's get HMR working for CSS at least.

Activating HMR: --hot
To activate it, stop the dev server and re-run it with a --hot flag... because HMR is the hotness... and also because it stands for hot module replacement:

 ./node_modules/.bin/webpack-dev-server --hot
Go refresh the browser to reset things. Now, make a change to our CSS and move back to the browser:

 82 lines  assets/css/main.scss
$ brand-primary: darken(#428bca, 2.5%); // #3885c7
Huh, it still refreshed. It's not working!

Webpack, AJAX and publicPath
Go to your browser's console, click the gear icon and check 'Preserve logs'. Thanks to this, the console will not clear logs when the page reloads.

Refresh once more... wait for the page to load... and then clear the console. Go make another change to the app, then move back. It says 'App updated', 'Recompiling', but then it refreshes the whole page!

These messages are coming from the HMR system. And check this out: it says 'Checking for updates on the server, and then you see a failed request to http://localhost:8000. Finally, it says 'Cannot find Update' and it refreshes the page.

What's going on? In addition to our built assets, the dev server - so localhost:8080 - serves a JSON file - this *.hot-update.json - with information about what was just updated.

So, why is that failing? Well, since our site actually lives at localhost:8000, when Webpack makes an AJAX call for this file, it makes it to localhost:8000 instead of localhost:8080. Internally, Webpack doesn't know that all of our assets are being hosted on some other host!

This is actually more important than just the dev server. Sometimes - like with a cool feature called code splitting that we'll see later - Webpack needs to make an AJAX call back to the server. In order to do this, it must know what the public URL is to your assets.

So... what's the problem? That is our public path, right? Not anymore! If you're using the webpack-dev-server - or if you're using a CDN on production - this publicPath must be absolute: it needs to include the host name where your assets are really stored. This is used for any Webpack AJAX calls and also to point to assets, like font files from inside CSS. That's why we saw some font file 404's in the last chapter.