The style-loader is the key: it embeds the CSS so that it's added to the page as a style tag. Basically, we need to replace the style-loader with something that, instead, outputs a real CSS file. That's exactly what the extract-text-webpack-plugin does!
Ignore the fallback key for a moment. Basically, the extract() function is a fancy way to prepend the loaders in use with a special extract-text-webpack-plugin loader. Thanks to this, CSS files will be processed through css-loader and then through this new extract text loader.
The styleLoader, which is set on fallback, is not used at all anymore.

Let's see what this does! Find your webpack tab and restart the dev server:

 ./node_modules/.bin/webpack-dev-server --hot
When it finishes, scroll up to the output. In addition to login.js, there is a login.css! And a layout.css and a rep_log.css. Our CSS is no longer packaged inside JavaScript: each entry now has its own CSS file!

 Tip

If a JavaScript entry file does not require any CSS, no .css file will be output by Webpack.