During development, we are not minifying our assets: yep, they're full of comments and spaces. And that's perfect for development! I like spaces! And those comments might be useful! Also... minification takes extra time, so we don't want to slow down our builds unnecessarily.
In other words, while developing, no minification! But for production, yes minification! That means that - somehow - we need to be able to pass a flag to webpack.config.js that tells it to compile in dev or production mode.

Using NODE_ENV
In Node applications, there's a standard way to signal the environment to your apps

Let's log that. Run webpack like normal:
./node_modules/.bin/webpack --watch
Ok! It prints undefined. So... how do we set that? On a UNIX system, prefix the command:

NODE_ENV=production ./node_modules/.bin/webpack --watch
Yes! It prints production. If you're on Windows, there's a library called cross-env that can help do this.

The point is: we can now send a flag into Webpack to tell it the environment.