Run: yarn add webpack-manifest-plugin --dev
Our script and link tags need to be updated for the new filenames. But... those filenames are now constantly changing!

Hmm. We need a map: a map to treasure. But also, another map that translates each original filename to the current, hashed filename. That map is called a manifest. webpack-manifest-plugin - can create it!