Read manifest.json file inside our app. Go back to our browser  and refresh. It works! Wait.. no it doesn't! What's up?
View the page source. Huh... it's still not using the hashed filename. When you use the json_manifest_path system, it literally takes whatever string is passed to asset() - so build/layout.css - and looks for that in manifest.json. Oh! But the key is just layout.css: there is no build/!