Add build/ prefix to entries in manifest.json file generated by webpack-manifest-plugin in order to be correctly read in our app.
Tip
In a future version of the plugin, you may need to also set publicPath to build/.

Re-run webpack:
yarn dev
Check out the manifest file. It's beautiful! And... refresh! Yes! It updates the path perfectly! Guys! We have built-in versioning! Whenever we update our files, Webpack will write a new filename and our app will use it.

This is great. But to take full advantage of it, you need to configure your web server to set a far-future Expires headers for everything in your /build directory... except for /build/static.