Go to the login page. If you type a really long username, we yell at you! Yes, we value brevity!
Look at the code behind this: assets/js/login.js. If the length is longer than 20, we add the warning:

 24 lines  assets/js/login.js
$ (document).ready(function() {
    $ ('.js-login-field-username').on('keydown', function(e) {
        if ($ usernameInput.val().length >= 20) {
            const $ warning = $ ('<div class='login-long-username-warning'>This is a really long username - are you sure that is right?</div>');
            $ usernameInput.before($ warning);
        }
    });
});
Code Splitting?
Guess what? Webpack comes with an absolutely killer feature... and I've been keeping it a secret! No more! For this to make sense, I want you to pretend that the code that adds the long username warning is really, really big. I literally mean, imagine the code inside the if statement is many lines... instead of just two.

In this situation, it's really wasteful to force the user to download all that extra JavaScript... just in case they ever type a long username!

In my perfect world, these two lines - or these many lines, in our imaginary situation - would only be downloaded by the user if and when they type a long username. Yes, I want to lazily load parts of our JavaScript!

With Webpack, this is possible! It's called 'code splitting'.