When you use import like this, it returns a Promise. Hey, we know about those! It means that we can say .then() and pass a callback.
It feels like... and well.. is an AJAX call! We're saying:

  Good afternoon Webpack! Could you please download the username_validation_error module via AJAX and then execute my callback when it's ready. Thank you!

It should work... but go look at the Webpack terminal. Ah! It's SO angry!

  Module build failed. Syntax error: import and export may only appear at the top level.

The dynamic import Proposal

Hmm. It's very unhappy about the import: it says that this is only allowed to live at the top of the file!

Here's the story: when you use import at the top of the file - like we've been doing until now - we're using a real, official ECMAScript feature - it's in ES6. But when you use import() like a function... well... that's not part of ECMAScript! Well, not yet. That functionality is just a proposal called 'dynamic import'.

The parse error comes from Babel: it tries to parse our code, but sees this code as invalid. And technically, it's right!