Run: yarn add babel-plugin-syntax-dynamic-import --dev
We need to teach Babel that this syntax is valid... but not to do anything with it. I mean, it should not have an error, and it should leave the import() function there so Webpack can parse it.

This is a plugin for Babel that makes it understand the dynamic import syntax.