Google for core-js. Ok, this is a library full of polyfills, for ES6, ES7 and other things. Ah, and here are the docs about Promise.
See the 'CommonJS entry points' part? This is telling us that if we need the Promise polyfill, we can just require core-js/library/es6/promise or core-js/library/fn/promise.

And now... re-dump our stats:
  NODE_ENV=production ./node_modules/.bin/webpack --json > stats.json

When it finishes, find the visualizer and refresh to reset it. Load the new stats.json. Oh, way better! jQuery now takes up 53%... not because it got bigger, but because our JavaScript donut is so much smaller. Then bootstrap-sass & sweetalert2. core-js went from 29.5% to only 8.1%.

Thank you visualizer! So, focus on your code! And then... come to the visualizer to make sure you're not doing anything crazy.